Added cooldown

+++ commands/advice.js
@@ -1,7 +1,17 @@
 /** @ignore */
 const snekfetch = require('snekfetch');
+const cooldown = new Set();
 
 exports.run = async (client,msg,args) => {
+    let toAdd = msg.author.id + msg.guild.id;
+    if (cooldown.has(toAdd)) {
+        return msg.reply('**[COOLDOWN]** Info command has **30 Seconds** Cooldown!');
+    }
+    cooldown.add(toAdd);
+
+    setTimeout(() => {
+        cooldown.delete(toAdd);
+    }, 30000);
     msg.delete(); //Deletes the message
 
     /**
@@ -11,7 +21,7 @@ exports.run = async (client,msg,args) => {
     */
     let r = await snekfetch.get("http://api.adviceslip.com/advice");
     let advice = JSON.parse(r.body).slip.advice;
-    msg.channel.send(advice).catch(e => logger.error(e))
+    msg.channel.send('**Advice: **' + advice).catch(e => logger.error(e))
 };
 
 exports.help = {
